完成角色修改，删除，管理员修改删除

+++ Application/Admin/Controller/RoleController.class.php
@@ -34,14 +34,54 @@ class RoleController extends IndexController
     		{
     			if($model->save() !== FALSE)
     			{
-    				$this->success('修改成功！', U('lst', array('p' => I('get.p', 1))));
-    				exit;
+    			    $rpModel = M('role_privilege');
+    			    $rpModel->where(array('role_id'=>I('post.id')))->delete();
+    			    $pri_id = I('post.pri_id');
+    			    if($pri_id)
+                    {
+                        foreach($pri_id as $v)
+                        {
+                            $res = $rpModel->add(
+                                array(
+                                    'pri_id' => $v,
+                                    'role_id' => I('post.id')
+                                )
+                            );
+                        }
+                    }
+                    if($res !== FALSE)
+                    {
+                        $this->success('修改成功！', U('lst', array('p' => I('get.p', 1))));
+                        exit;
+                    }
     			}
     		}
     		$this->error($model->getError());
     	}
     	$model = M('Role');
     	$data = $model->find($id);
+    	$rpModel = M('role_privilege');
+    	$rpData = $rpModel->alias('a')->where(array('role_id'=>$id))->join('LEFT JOIN rc_privilege b ON a.pri_id = b.id')->select();
+    	$priModel = D('Admin/Privilege');
+    	$priData = $priModel->getTree();
+
+    	foreach($priData as $k => $v)
+        {
+            foreach($rpData as $k0 => $v0)
+            {
+                if($v0['pri_id'] == $v['id'])
+                {
+                    $priData[$k]['is_selected'] = 1;
+                    break;
+                }
+                else
+                {
+                    $priData[$k]['is_selected'] = 0;
+                }
+            }
+        }
+    	$this->assign('rpData',$rpData);
+    	$this->assign('priData',$priData);
     	$this->assign('data', $data);
 
 		$this->setPageBtn('修改角色', '角色列表', U('lst?p='.I('get.p')));
@@ -50,10 +90,22 @@ class RoleController extends IndexController
     public function delete()
     {
     	$model = D('Admin/Role');
+    	$arModel = M('admin_role');
+    	$is_exist = $arModel->where(array('role_id'=>I('get.id')))->find();
+    	if(!empty($is_exist))
+        {
+            $this->error('存在后台管理员属于该角色，不能删除！');
+            exit;
+        }
     	if($model->delete(I('get.id', 0)) !== FALSE)
     	{
-    		$this->success('删除成功！', U('lst', array('p' => I('get.p', 1))));
-    		exit;
+    	    $rpModel = M('role_privilege');
+    	    $delRes = $rpModel->where(array('role_id'=>I('get.id')))->delete();
+    	    if($delRes)
+            {
+                $this->success('删除成功！', U('lst', array('p' => I('get.p', 1))));
+                exit;
+            }
     	}
     	else 
     	{
